move tests for FormatsOptions to their own class
