+++ components/formats-api/test/loci/formats/utests/testng-template.xml
@@ -36,6 +36,11 @@
         <class name="loci.formats.utests.FormatToolsTest"/>
       </classes>
     </test>
+    <test name="FormatsOptions">
+      <classes>
+        <class name="loci.formats.utests.FormatsOptionsTest"/>
+      </classes>
+    </test>
     <test name="DefaultMetadataOptions">
       <classes>
         <class name="loci.formats.utests.DefaultMetadataOptionsTest"/>
